requirements.txt
absl-py==2.0.0
aio-pika==9.4.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiormq==6.8.1
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.32.0
azure-storage-blob==12.16.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blis==1.2.0
boto3==1.35.99
botocore==1.35.99
CacheControl==0.14.2
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==2.0.0
comm==0.2.2
confection==0.1.5
confluent-kafka==2.8.0
contourpy==1.3.1
croniter==6.0.0
cryptography==44.0.1
cvg-python-sdk==0.5.1
cycler==0.12.1
cymem==2.0.11
dask==2024.7.1
dataclasses-json==0.6.7
debtcollector==3.0.0
debugpy==1.8.12
decorator==5.1.1
demoji==1.1.0
Deprecated==1.2.18
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
dogpile.cache==1.3.4
environs==9.5.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
eventlet==0.39.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
Faker==26.3.0
fasteners==0.19
fbmessenger==6.0.0
filelock==3.17.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
futurist==3.1.0
gast==0.6.0
google-api-core==2.8.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.56.1
greenlet==3.1.1
grpcio==1.60.0
grpcio-health-checking==1.59.5
grpcio-tools==1.60.0
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
hvac==1.2.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.1.3
ipykernel==6.29.5
ipython==8.32.0
iso8601==2.1.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
keras==2.14.0
keystoneauth1==5.10.0
keystonemiddleware==10.9.0
kiwisolver==1.4.8
kombu==5.4.2
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langcodes==3.5.0
langsmith==0.1.147
language_data==1.3.0
libclang==18.1.1
litellm==1.52.16
load-dotenv==0.1.0
locket==1.0.0
logutils==0.3.5
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.7.5
matplotlib-inline==0.1.7
mattermostwrapper==2.2
mdurl==0.1.2
minio==7.2.15
mistral==19.0.0
mistral-lib==3.3.1
mistralai==1.5.0
ml-dtypes==0.2.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.55.3
openpyxl==3.1.5
opentelemetry-api==1.16.0
opentelemetry-exporter-jaeger==1.16.0
opentelemetry-exporter-jaeger-proto-grpc==1.16.0
opentelemetry-exporter-jaeger-thrift==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opt_einsum==3.4.0
orjson==3.10.15
os-service-types==1.7.0
oslo.cache==3.10.1
oslo.concurrency==7.1.0
oslo.config==9.7.1
oslo.context==5.7.1
oslo.db==17.2.1
oslo.i18n==6.5.1
oslo.log==7.1.0
oslo.messaging==16.1.0
oslo.metrics==0.11.0
oslo.middleware==6.3.1
oslo.policy==4.5.1
oslo.serialization==5.7.0
oslo.service==4.1.1
oslo.utils==8.2.0
osprofiler==4.2.0
packaging==23.2
pamqp==3.3.0
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
partd==1.4.2
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.1
pecan==1.5.1
pep440-version-utils==1.1.1
pexpect==4.9.0
phonenumbers==8.13.55
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
portalocker==2.10.1
preshed==3.0.9
presidio-analyzer==2.2.33
presidio_anonymizer==2.2.357
prettytable==3.15.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycadf==4.0.1
pycountry==22.3.5
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.2
Pygments==2.19.1
pyhcl==0.4.5
PyJWT==2.10.1
pykwalify==1.8.0
pymilvus==2.4.1
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
pypred==0.4.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-keycloak==3.12.0
python-keystoneclient==5.5.0
python-socketio==5.12.1
pytz==2022.7.1
PyYAML==6.0.2
pyzmq==26.2.1
qdrant-client==1.13.2
questionary==2.0.1
randomname==0.2.1
rasa-pro==3.11.4
rasa-sdk==3.11.0
redis==5.2.1
referencing==0.36.2
regex==2022.10.31
repoze.lru==0.7
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rocketchat-API==1.30.0
Routes==2.5.1
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
safetensors==0.4.5
sanic==22.12.0
Sanic-Cors==2.2.0
sanic-jwt==1.8.0
sanic-routing==22.8.0
scikit-learn==1.5.2
scipy==1.13.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
simplegeneric==0.8.1
six==1.17.0
sklearn-crfsuite==0.3.6
skops==0.10.0
slack_sdk==3.27.2
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
statsd==4.0.1
stevedore==5.4.1
structlog==23.1.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.5
tenacity==8.4.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-estimator==2.14.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-text==2.14.0
termcolor==2.5.0
terminaltables==3.1.10
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
thinc==8.3.4
threadpoolctl==3.5.0
thrift==0.21.0
tiktoken==0.7.0
tldextract==5.1.3
tokenizers==0.21.0
toolz==1.0.0
tooz==6.3.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==8.4.0
typer==0.15.1
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
ujson==5.10.0
urllib3==2.3.0
uv==0.6.0
uvloop==0.21.0
vine==5.1.0
voluptuous==0.15.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webexteamssdk==1.6.1
WebOb==1.8.9
websockets==10.4
Werkzeug==3.1.3
wrapt==1.14.1
WSME==0.12.1
wsproto==1.2.0
yappi==1.6.10
yaql==3.0.0
yarl==1.18.3
zipp==3.21.0
